// src/app/page/master/jenis-petugas-pelaksana/jenis-petugas-pelaksana.interface.ts
    import {EventEmitter} from '@angular/core';

    export interface JenisPetugasPelaksana {
		kdProfile?: string;
		jenisPajak?: string;
		kdJenisPajakHead?: number;
		deskripsi?: string;
		kdDepartemen?: number;
		reportDisplay?: string;
		kodeExternal?: string;
		namaExternal?: string;
		tampilanLaporan?: string;
		noRec?: string;
		kode?:any;
    }